python: Move the ACS types in dazl.model.core closer to their acutal usage.
