+++ python/dazl/client/state.py
@@ -3,18 +3,60 @@
 
 from asyncio import Future, sleep
 from collections import defaultdict
-from typing import Awaitable, Collection, Dict, List, Optional, Union, cast
+from dataclasses import dataclass
+import datetime
+from typing import Awaitable, Collection, Dict, List, Optional, Tuple, Union, cast
 import warnings
 
 from ..damlast.daml_lf_1 import TypeConName
 from ..damlast.protocols import SymbolLookup
-from ..model.core import ContractContextualData, ContractContextualDataCollection, ContractsState
 from ..model.reading import ContractArchiveEvent, ContractCreateEvent
-from ..prim import ContractId
+from ..prim import ContractData, ContractId
 from ..query import ContractMatch, is_match
 from ..scheduler import Invoker
 from ..util.asyncio_util import await_then
 from .errors import UnknownTemplateWarning
+
+__all__ = [
+    "ContractsState",
+    "ContractsHistoricalState",
+    "ContractContextualData",
+    "ContractContextualDataCollection",
+    "ActiveContractSet",
+    "PendingQuery",
+]
+
+ContractsState = Dict[ContractId, ContractData]
+ContractsHistoricalState = Dict[ContractId, Tuple[ContractData, bool]]
+
+
+class ContractContextualDataCollection(tuple):
+    def __getitem__(self, index: "Union[int, str, ContractId]"):
+        if index is None:
+            raise ValueError("the index cannot be None")
+        elif isinstance(index, int):
+            return tuple.__getitem__(self, index)
+        elif isinstance(index, str):
+            for cxd in self:
+                if cxd.cid.contract_id == index:
+                    return cxd
+            raise KeyError(index)
+        elif isinstance(index, ContractId):
+            for cxd in self:
+                if cxd.cid == index:
+                    return cxd
+            raise KeyError(index)
+        else:
+            raise TypeError("cannot index into a ContractContextualDataCollection with {index!r}")
+
+
+@dataclass(frozen=True)
+class ContractContextualData:
+    cid: "ContractId"
+    cdata: "Optional[ContractData]"
+    effective_at: datetime
+    archived_at: "Optional[datetime]"
+    active: bool
 
 
 class ActiveContractSet:
